Default app selection, streamlined downloads
The UI parts and affordances related to setting of default app can now
be turned off.
Refactored code so that downloads of icons/app types are singular (for
example, to avoid multi-download of same app icon when an app like
Safari is registered with multiple app types like Browser and Twitter).